requirements.txt
pandas
numpy
requests
redis
redisvl
python-dotenv
sentence-transformers
pyyaml